+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.0.12] - 2024-08-31
+- only run multi-threaded runtime when there is more than one thread available
+- stop using custom allocator
+- ability to add to startup and run as a daemon
+
+
 ## [0.0.11] - 2024-08-5
 - only run multi-threaded runtime when there is more than one thread available
 - stop using custom allocator